Client et serveur: * Modification de noms de classes
Serveur:
* Nouvelle classe "Clients" pour centraliser les fonctions en lien avec
les clients.

+++ Client/Client_Synergie/Reseau/Paquets/paquetchangernom.cpp
@@ -1,7 +1,7 @@
 #include "paquetchangernom.h"
 
 PaquetChangerNom::PaquetChangerNom(QString nom) :
-    BasePaquet((qint8)1)
+    BasePaquet(1)
 {
     *m_Stream << nom;
 }
