+++ Client/Client_Synergie/Reseau/Paquets/paqueteffacementtexte.cpp
@@ -1,7 +1,7 @@
 #include "paqueteffacementtexte.h"
 
 PaquetEffacementTexte::PaquetEffacementTexte(int id, int position, int longeur) :
-    BasePaquet((qint8)4)
+    BasePaquet(4)
 {
     *m_Stream << id;
     *m_Stream << position;
