+++ Client/Client_Synergie/Reseau/Paquets/paquetenvoichat.cpp
@@ -1,7 +1,7 @@
 #include "paquetenvoichat.h"
 
 PaquetEnvoiChat::PaquetEnvoiChat(QString message):
-    BasePaquet((qint8)10)
+    BasePaquet(10)
 {
     *m_Stream << message;
 }
