+++ Client/Client_Synergie/Reseau/Paquets/paquetinsertiontexte.cpp
@@ -1,7 +1,7 @@
 #include "paquetinsertiontexte.h"
 
 PaquetInsertionTexte::PaquetInsertionTexte(int id, int pos, QString texte):
-    BasePaquet((qint8)2)
+    BasePaquet(2)
 {
     *m_Stream << id;
     *m_Stream << pos;
