+++ Client/Client_Synergie/Reseau/Paquets/paquetouvrirfeuille.cpp
@@ -1,7 +1,7 @@
 #include "paquetouvrirfeuille.h"
 
 PaquetOuvrirFeuille::PaquetOuvrirFeuille(int id) :
-    BasePaquet((qint8)7)
+    BasePaquet(7)
 {
     *m_Stream << id;
 }
