+++ Client/Client_Synergie/Reseau/Paquets/paquetreceptiondonnees.cpp
@@ -1,7 +1,7 @@
 #include "paquetreceptiondonnees.h"
 
 PaquetReceptionDonnees::PaquetReceptionDonnees(int id) :
-    BasePaquet((qint8)10)
+    BasePaquet(10)
 {
     *m_Stream << id;
 }
